refactor(python)!: flatten tracking.config and endgame.config into parent modules
The config sub-submodules were awkward (bertini.endgame.config.Endgame was
a config class confusingly named just 'Endgame'). Config classes now live
directly in bertini.tracking and bertini.endgame, with Config-suffixed names:

- endgame.config.Endgame   -> endgame.EndgameConfig
- endgame.config.Security  -> endgame.SecurityConfig
- endgame.config.{PowerSeries,Cauchy}Config -> endgame.* (unchanged names)
- tracking.config.*        -> tracking.* (unchanged names)

BREAKING CHANGE: bertini.tracking.config and bertini.endgame.config no
longer exist; import the config classes from the parent modules.
