+++ python/bertini/endgame/__init__.py
@@ -40,7 +40,7 @@ There are two basic flavors of endgame implemented:
 1. Power Series, commonly written PS or PSEG
 2. Cauchy
 
-Both estimate the cycle number and use it to compute a root at a time which is never tracked to.  PSEG uses Hermite interpolation and extrapolation, and Cauchy uses loops around the target time coupled with the `Cauchy integral formula <https://en.wikipedia.org/wiki/Cauchy%27s_integral_formula>`_.  Both continue until two successive approximations of the root match to a given tolerance (:py:attr:`bertini.endgame.config.Endgame.final_tolerance`).
+Both estimate the cycle number and use it to compute a root at a time which is never tracked to.  PSEG uses Hermite interpolation and extrapolation, and Cauchy uses loops around the target time coupled with the `Cauchy integral formula <https://en.wikipedia.org/wiki/Cauchy%27s_integral_formula>`_.  Both continue until two successive approximations of the root match to a given tolerance (:py:attr:`bertini.endgame.EndgameConfig.final_tolerance`).
 
 The implementations of the endgames go with a particular tracker, hence there are six provided endgame types.  Choose the one that goes with your selected tracker type.  Adaptive Multiple Precision is a good choice.
 
@@ -70,7 +70,17 @@ Fixed Multiple Precision  Endgames
 
 import bertini._pybertini.endgame as _pybe
 
-from bertini._pybertini.endgame import config, observers
+from bertini._pybertini.endgame import observers
+from bertini._pybertini.endgame import (
+    CauchyConfig,
+    EndgameConfig,
+    PowerSeriesConfig,
+    SecurityConfig,
+)
+
+# configs gain update()/to_dict()/from_dict()/repr/eq
+from ..config import enhance_all
+enhance_all(_pybe)
 
 
 class _EndgameBase:
@@ -121,10 +131,13 @@ class FixedMultiplePSEG(_EndgameBase):
 __all__ = [
     'AMPCauchyEG',
     'AMPPSEG',
+    'CauchyConfig',
+    'EndgameConfig',
     'FixedDoubleCauchyEG',
     'FixedDoublePSEG',
     'FixedMultipleCauchyEG',
     'FixedMultiplePSEG',
-    'config',
+    'PowerSeriesConfig',
+    'SecurityConfig',
     'observers',
 ]
