+++ python/test/tracking/predictor_test.py
@@ -41,7 +41,7 @@ from bertini import System, VariableGroup
 from bertini.function_tree.symbol import Variable
 from bertini.function_tree import *
 from bertini.tracking import AMPTracker, Predictor, SuccessCode
-from bertini.tracking.config import SteppingConfig, NewtonConfig, amp_config_from
+from bertini.tracking import SteppingConfig, NewtonConfig, amp_config_from
 
 import bertini.multiprec as mp
 from bertini.multiprec import Complex as mpfr_complex
