+++ python_bindings/src/endgame_export.cpp
@@ -113,17 +113,8 @@ namespace bertini{
 
 		void ExportEndgameSettings()
 		{
-			scope current_scope;
-			std::string new_submodule_name(extract<const char*>(current_scope.attr("__name__")));
-			new_submodule_name.append(".config");
-			object new_submodule(borrowed(PyImport_AddModule(new_submodule_name.c_str())));
-			current_scope.attr("config") = new_submodule;
-			
-
-			scope new_submodule_scope = new_submodule;
-			new_submodule_scope.attr("__doc__") = "Endgame configuration structs.";
-
-			class_<endgame::EndgameConfig>("Endgame","Generic endgame settings.  Number of sample points, etc.  Note that some of its configs are rational numbers",init<>())
+			// config classes live directly in the endgame module, with Config-suffixed names
+			class_<endgame::EndgameConfig>("EndgameConfig","Generic endgame settings.  Number of sample points, etc.  Note that some of its configs are rational numbers",init<>())
 				.def_readwrite("sample_point_refinement_factor", &endgame::EndgameConfig::sample_point_refinement_factor, "Extra amount of tolerance for refining before computing the final approximation, during endgame.")
 				.def_readwrite("num_sample_points", &endgame::EndgameConfig::num_sample_points,"The number of points to use for extrapolant calculation.  In the Power Series Endgame, the is the number of geometrically spaces points on the path.  For Cauchy, this is the number of points on each circle tracked around the target time value.")
 				.def_readwrite("min_track_time", &endgame::EndgameConfig::min_track_time,"The minimum distance from the target time to track to.  Decreasing this may help failing runs succeed, or maybe not, because you are, after all, tracking toward a singularity.")
@@ -133,7 +124,7 @@ namespace bertini{
 				;
 
 
-			class_<endgame::SecurityConfig>("Security","Security settings for endgames.  Control things like truncation because estimated root is near infinity",init<>())
+			class_<endgame::SecurityConfig>("SecurityConfig","Security settings for endgames.  Control things like truncation because estimated root is near infinity",init<>())
 				.def_readwrite("level", &endgame::SecurityConfig::level,"Turns on or off truncation of paths going to infinity during the endgame.  0 is off, 1 is on.")
 				.def_readwrite("max_norm", &endgame::SecurityConfig::max_norm,"If on, the norm at which to truncate a path.")
 				;
