+++ python_bindings/src/tracker_export.cpp
@@ -39,7 +39,7 @@ namespace bertini{
 		void TrackerVisitor<TrackerT>::visit(PyClass& cl) const
 		{
 			cl
-			.def("setup", &TrackerT::Setup, (arg("predictor"), arg("tolerance"), arg("truncation"), arg("stepping"),arg("newton")), "Set values for the internal configuration of the tracker.  tolerance and truncation are both real doubles.  predictor is a valid value for predictor choice.  stepping and newton are the config structs from bertini.tracking.config.")
+			.def("setup", &TrackerT::Setup, (arg("predictor"), arg("tolerance"), arg("truncation"), arg("stepping"),arg("newton")), "Set values for the internal configuration of the tracker.  tolerance and truncation are both real doubles.  predictor is a valid value for predictor choice.  stepping and newton are the config structs from bertini.tracking.")
 
 			.def("track_path", &track_path_wrap, 
 				 (arg("self"),arg("result"), "start_time", "end_time", "start_point"), 
@@ -244,58 +244,49 @@ namespace bertini{
 				.value("FailedToSelectPrecisionAndStepsize", SuccessCode::FailedToSelectPrecisionAndStepsize)
 				;
 			
-			{ // enter a scope for config types
-				scope current_scope;
-				std::string new_submodule_name(extract<const char*>(current_scope.attr("__name__")));
-				new_submodule_name.append(".config");
-				object new_submodule(borrowed(PyImport_AddModule(new_submodule_name.c_str())));
-				current_scope.attr("config") = new_submodule;
+			// config classes live directly in the tracking module
 
-				scope new_submodule_scope = new_submodule;
+			// class_<config::Tolerances<double>>("Tolerances_d",init<>())
+			// 	.def(TolerancesVisitor<double>());
 
+			// class_<config::Tolerances<mpfr_float>>("Tolerances_mp",init<>())
+			// 	.def(TolerancesVisitor<mpfr_float>());
 
-				// class_<config::Tolerances<double>>("Tolerances_d",init<>())
-				// 	.def(TolerancesVisitor<double>());
-
-				// class_<config::Tolerances<mpfr_float>>("Tolerances_mp",init<>())
-				// 	.def(TolerancesVisitor<mpfr_float>());
-
-				class_<SteppingConfig, std::shared_ptr<SteppingConfig> >("SteppingConfig", init<>())
-					.def(SteppingVisitor<double>())
-					;
-				
-				
-				class_<NewtonConfig, std::shared_ptr<NewtonConfig> >("NewtonConfig", init<>())
-					.def_readwrite("max_num_newton_iterations", &NewtonConfig::max_num_newton_iterations)
-					.def_readwrite("min_num_newton_iterations", &NewtonConfig::min_num_newton_iterations)
-					;
-				
-				
-				class_<FixedPrecisionConfig, std::shared_ptr<FixedPrecisionConfig> >("FixedPrecisionConfig", init<System const&>());
+			class_<SteppingConfig, std::shared_ptr<SteppingConfig> >("SteppingConfig", init<>())
+				.def(SteppingVisitor<double>())
+				;
 
 
-				
-				
-				class_<AdaptiveMultiplePrecisionConfig, std::shared_ptr<AdaptiveMultiplePrecisionConfig> >("AMPConfig", init<>())
-					.def(init<System const&>())
-					.def("set_amp_config_from", &AdaptiveMultiplePrecisionConfig::SetAMPConfigFrom)
-					.def("set_phi_psi_from_bounds", &AdaptiveMultiplePrecisionConfig::SetPhiPsiFromBounds)
-					.def("set_bounds_and_epsilon_from", &AdaptiveMultiplePrecisionConfig::SetBoundsAndEpsilonFrom)
-					.def_readwrite("coefficient_bound", &AdaptiveMultiplePrecisionConfig::coefficient_bound)
-					.def_readwrite("degree_bound", &AdaptiveMultiplePrecisionConfig::degree_bound)
-					.def_readwrite("epsilon", &AdaptiveMultiplePrecisionConfig::epsilon)
-					.def_readwrite("phi", &AdaptiveMultiplePrecisionConfig::Phi)
-					.def_readwrite("psi", &AdaptiveMultiplePrecisionConfig::Psi)
-					.def_readwrite("safety_digits_1", &AdaptiveMultiplePrecisionConfig::safety_digits_1)
-					.def_readwrite("safety_digits_2", &AdaptiveMultiplePrecisionConfig::safety_digits_2)
-					.def_readwrite("maximum_precision", &AdaptiveMultiplePrecisionConfig::maximum_precision)
-					.def_readwrite("consecutive_successful_steps_before_precision_decrease", &AdaptiveMultiplePrecisionConfig::consecutive_successful_steps_before_precision_decrease)
-					.def_readwrite("max_num_precision_decreases", &AdaptiveMultiplePrecisionConfig::max_num_precision_decreases)
-					.def_readwrite("coefficient_bound", &AdaptiveMultiplePrecisionConfig::coefficient_bound)
-					;
-				
-				def("amp_config_from", &AMPConfigFrom, "make an AMPConfig from a System with generated settings for system-specific things, and default settings otherwise (such as safety digits).");
-			}
+			class_<NewtonConfig, std::shared_ptr<NewtonConfig> >("NewtonConfig", init<>())
+				.def_readwrite("max_num_newton_iterations", &NewtonConfig::max_num_newton_iterations)
+				.def_readwrite("min_num_newton_iterations", &NewtonConfig::min_num_newton_iterations)
+				;
+
+
+			class_<FixedPrecisionConfig, std::shared_ptr<FixedPrecisionConfig> >("FixedPrecisionConfig", init<System const&>());
+
+
+
+
+			class_<AdaptiveMultiplePrecisionConfig, std::shared_ptr<AdaptiveMultiplePrecisionConfig> >("AMPConfig", init<>())
+				.def(init<System const&>())
+				.def("set_amp_config_from", &AdaptiveMultiplePrecisionConfig::SetAMPConfigFrom)
+				.def("set_phi_psi_from_bounds", &AdaptiveMultiplePrecisionConfig::SetPhiPsiFromBounds)
+				.def("set_bounds_and_epsilon_from", &AdaptiveMultiplePrecisionConfig::SetBoundsAndEpsilonFrom)
+				.def_readwrite("coefficient_bound", &AdaptiveMultiplePrecisionConfig::coefficient_bound)
+				.def_readwrite("degree_bound", &AdaptiveMultiplePrecisionConfig::degree_bound)
+				.def_readwrite("epsilon", &AdaptiveMultiplePrecisionConfig::epsilon)
+				.def_readwrite("phi", &AdaptiveMultiplePrecisionConfig::Phi)
+				.def_readwrite("psi", &AdaptiveMultiplePrecisionConfig::Psi)
+				.def_readwrite("safety_digits_1", &AdaptiveMultiplePrecisionConfig::safety_digits_1)
+				.def_readwrite("safety_digits_2", &AdaptiveMultiplePrecisionConfig::safety_digits_2)
+				.def_readwrite("maximum_precision", &AdaptiveMultiplePrecisionConfig::maximum_precision)
+				.def_readwrite("consecutive_successful_steps_before_precision_decrease", &AdaptiveMultiplePrecisionConfig::consecutive_successful_steps_before_precision_decrease)
+				.def_readwrite("max_num_precision_decreases", &AdaptiveMultiplePrecisionConfig::max_num_precision_decreases)
+				.def_readwrite("coefficient_bound", &AdaptiveMultiplePrecisionConfig::coefficient_bound)
+				;
+
+			def("amp_config_from", &AMPConfigFrom, "make an AMPConfig from a System with generated settings for system-specific things, and default settings otherwise (such as safety digits).");
 			
 		}
